INW-154 removed flipview and shimmer_recycler modules, files moved to designsystem
