+++ Android/app/build.gradle
@@ -130,8 +130,6 @@ protobuf {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
 
-    implementation project(path: ':flipview')
-    implementation project(path: ':shimmer_recycler')
     implementation project(path: ':designcorelib')
     implementation project(path: ':networklib')
     implementation project(path: ':corelib')
