+++ Android/designcorelib/src/main/res/anim/grow_from_middle_x_axis.xml
