+++ Android/designcorelib/src/main/res/anim/scale_down.xml
