+++ Android/designcorelib/src/main/res/anim/scale_up.xml
