+++ Android/designcorelib/src/main/res/anim/shrink_to_middle_x_axis.xml
