+++ Android/designcorelib/src/main/res/drawable-hdpi/ic_action_done.png
