+++ Android/designcorelib/src/main/res/drawable-mdpi/ic_action_done.png
