+++ Android/designcorelib/src/main/res/drawable-xhdpi/ic_action_done.png
