+++ Android/designcorelib/src/main/res/drawable-xxhdpi/ic_action_done.png
