+++ Android/designcorelib/src/main/res/drawable-xxxhdpi/ic_action_done.png
