+++ Android/designcorelib/src/main/res/layout/flipview_front.xml
