+++ Android/designcorelib/src/main/res/layout/flipview_rear.xml
