+++ Android/designcorelib/src/main/res/layout/layout_sample_view.xml
