+++ Android/designcorelib/src/main/res/layout/viewholder_shimmer.xml
