+++ Android/designcorelib/src/main/res/values/attrs.xml
