+++ Android/designcorelib/src/main/res/values/attrs_shimmer_recycler.xml
