+++ Android/designcorelib/src/main/res/values/colors_shimmer_recycler.xml
