--- Android/flipview/.gitignore
@@ -1,1 +1,0 @@
-/build
