--- Android/flipview/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="ru.inwords.flipview"/>
