--- Android/shimmer_recycler/.gitignore
@@ -1,1 +1,0 @@
-/build
