--- Android/shimmer_recycler/build.gradle
@@ -1,31 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion targetSdk
-
-    defaultConfig {
-        minSdkVersion minSdk
-        targetSdkVersion targetSdk
-        versionCode 1
-        versionName "1.0"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation "androidx.recyclerview:recyclerview:$versions.RECYCLER"
-    implementation "com.facebook.shimmer:shimmer:$versions.SHIMMER"
-}
