--- Android/shimmer_recycler/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="ru.inwords.shimmer_recycler" />
